SW-8698 Replace full viewer with wrapper over web-components version
